+++ src/main/java/janggi/piece/limit/Guard.java
@@ -1,18 +1,19 @@
-package janggi.piece;
+package janggi.piece.limit;
 
 import janggi.board.Position;
 import janggi.move.Direction;
 import janggi.move.Route;
+import janggi.piece.PieceType;
+import janggi.piece.Side;
 import java.util.List;
 
-public class King extends LimitMovable {
+public class Guard extends LimitMovable {
 
-    public King(final Side side) {
+    public Guard(final Side side) {
         super(side);
     }
 
-    @Override
-    public List<Route> computeCandidatePositions(final Position position) {
+    public List<Route> computeCandidatePositions(Position position) {
         return List.of(
                 createRoute(position, Direction.UP),
                 createRoute(position, Direction.RIGHT),
@@ -29,6 +30,6 @@ public class King extends LimitMovable {
 
     @Override
     public PieceType getType() {
-        return PieceType.KING;
+        return PieceType.GUARD;
     }
 }
