+++ src/main/java/janggi/piece/limit/Soldier.java
@@ -1,11 +1,12 @@
-package janggi.piece;
+package janggi.piece.limit;
 
 import janggi.board.Position;
 import janggi.move.Direction;
 import janggi.move.Route;
 
+import janggi.piece.PieceType;
+import janggi.piece.Side;
 import java.util.List;
-import java.util.Map;
 
 public class Soldier extends LimitMovable {
