+++ src/test/java/janggi/board/JanggiBoardTest.java
@@ -2,6 +2,14 @@ package janggi.board;
 
 import janggi.piece.Side;
 import janggi.piece.*;
+import janggi.piece.limit.Elephant;
+import janggi.piece.Empty;
+import janggi.piece.limit.Guard;
+import janggi.piece.limit.Horse;
+import janggi.piece.limit.King;
+import janggi.piece.limit.Soldier;
+import janggi.piece.unlimit.Cannon;
+import janggi.piece.unlimit.Chariot;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
