+++ src/test/java/janggi/piece/limit/KingTest.java
@@ -1,28 +1,29 @@
-package janggi.piece;
+package janggi.piece.limit;
 
 import janggi.board.Position;
 import janggi.move.Route;
+import janggi.piece.Side;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.*;
 
-class GuardTest {
+class KingTest {
 
     @Test
     @DisplayName("현재 위치를 받아와 움직일 수 있는 위치 후보군을 반환한다")
     void computeCandidatePositions() {
 
-        Guard guard = new Guard(Side.CHO);
-        Position currentPosition = new Position(3, 9);
-        List<Route> candidatePositions = guard.computeCandidatePositions(currentPosition);
+        King king = new King(Side.CHO);
+        Position currentPosition = new Position(4, 8);
+        List<Route> candidatePositions = king.computeCandidatePositions(currentPosition);
 
         assertAll(
                 () -> assertThat(candidatePositions).hasSize(4),
-                () -> assertThat(candidatePositions.getLast().getLastPosition()).isEqualTo(new Position(4, 9))
+                () -> assertThat(candidatePositions.getLast().getLastPosition()).isEqualTo(new Position(5, 8))
         );
     }
