+++ src/test/java/janggi/piece/unlimit/ChariotTest.java
@@ -1,7 +1,8 @@
-package janggi.piece;
+package janggi.piece.unlimit;
 
 import janggi.board.Position;
 import janggi.move.Route;
+import janggi.piece.Side;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
